requirements.txt
aiohttp==3.8.4
albumentations==1.3.1
docstring-parser==0.15
ftfy==6.1.1
lightning==2.0.4
lightning-bolts==0.7.0
lightning-flash==0.8.1.post0
pip-chill==1.0.3
pystiche==1.0.1
regex==2023.6.3
sahi==0.10.8
segmentation-models-pytorch==0.3.3
torchaudio==2.1.0.dev20230701
